long removed

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,8 +6,8 @@
  */
 int main(void)
 {
-	long int sum;
-	long int first, second;
+	int sum;
+	int first, second;
 	int count;
 
 	sum = 0;
